renamed numbers.py -> number_management.py to avoid namespace clash (#247)
